bug fix (wrong types)

+++ inc/io.h
@@ -12,8 +12,8 @@
 
 using namespace std;
 
-void get_initial_values(string *path, uint64_t *steps, double *dt, uint32_t *save_interval, uint32_t *ignore_bodies,
-                        float *G);
+void get_initial_values(string *path, unsigned long long int *steps, double *dt, unsigned long int *save_interval,
+                        unsigned long int *ignore_bodies, float *G);
 
 void write_file(vector<Body> bodies, char filename[], double dt, double t);
