+++ src/io.cpp
@@ -9,8 +9,8 @@
 #include <sstream>
 #include <string>
 
-void get_initial_values(string *path, uint64_t *steps, double *dt, uint32_t *save_interval, uint32_t *ignore_bodies,
-                        float *G) {
+void get_initial_values(string *path, unsigned long long int *steps, double *dt, unsigned long int *save_interval,
+                        unsigned long int *ignore_bodies,  float *G) {
     string input_path = "../input/input.conf";
 
     // std::ifstream is RAII, i.e. no need to call close
